requirements.txt
pyqtgraph
pyserial
PySide6
matplotlib
numpy
scipy
pyautogui
pydirectinput